# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Open Live API
  version: 1.0.0
  description: |
    REST API for the Open Live broadcast production platform.

    **Authentication**: Pass an API key via `Authorization: Bearer <key>` header or `?key=<key>` query parameter.
    The following paths are exempt from authentication: `/health`, `/ready`, `/api/v1/status`, `/api/v1/server-info`.
servers:
  - url: /
security:
  - bearerAuth: []
  - apiKeyQuery: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyQuery:
      type: apiKey
      in: query
      name: key
  schemas:
...
    StreamType:
      type: string
      enum: [srt, efp, whip, html, clip]
      description: >-
        Source stream kind. `clip` (epic #206) is a cue-then-play file/story
        source whose typed `ClipReference` is stored serialized in the source's
        `address` field. Clients must ignore unknown enum values (forward-compat).
...
    SourceStatus:
      type: string
      enum: [active, inactive]
...
    Source:
      type: object
      required: [id, name, address, streamType, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000
          description: SRT latency in milliseconds
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SourceInput:
      type: object
      required: [name, address, streamType]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
          default: inactive
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000
    SourcePatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        streamType:
          $ref: '#/components/schemas/StreamType'
        status:
          $ref: '#/components/schemas/SourceStatus'
        liveCamera:
          type: boolean
        latency:
          type: integer
          minimum: 20
          maximum: 8000

    # ── Graphic ─────────────────────────────────────────────────────────────────
...
    Error:
      type: object
      properties:
        error:
          type: string
        statusCode:
          type: integer

    # ── Guest calling (epic #208, issue #299) ─────────────────────────────────────
...
paths:
...
  /api/v1/sources:
    get:
      summary: List sources
      operationId: listSources
      tags: [Sources]
      responses:
        '200':
          description: Array of sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Source'
    post:
      summary: Create source
      operationId: createSource
      tags: [Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceInput'
      responses:
        '201':
          description: Created source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Hostless SRT listener port is outside this instance's leased range (see /api/v1/server-info)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >-
            Hostless SRT listener port is already held by another source or output, or no
            port is free in the range for a port-0 assignment request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: SRT port range not yet allocated from Strom — retry shortly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/sources/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get source
      operationId: getSource
      tags: [Sources]
      responses:
        '200':
          description: Source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Not found
    patch:
      summary: Update source
      operationId: updateSource
      tags: [Sources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcePatch'
      responses:
        '200':
          description: Updated source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '404':
          description: Not found
        '422':
          description: Hostless SRT listener port is outside this instance's leased range (see /api/v1/server-info)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: >-
            Hostless SRT listener port is already held by another source or output, or no
            port is free in the range for a port-0 assignment request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: SRT port range not yet allocated from Strom — retry shortly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete source
      operationId: deleteSource
      tags: [Sources]
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found

  # ── Graphics ─────────────────────────────────────────────────────────────────
...