# src/routes/sources.ts
import type { FastifyPluginAsync } from 'fastify';
import { randomUUID } from 'crypto';
import { z } from 'zod';
import { getSourcesDb, getGatewaysDb, getDb } from '../db/index.js';
import type { SourceDoc, ProductionDoc } from '../db/types.js';
import { updateProductionDoc } from './productions.js';
import { graphicUrl, srtUrl } from '../lib/url-validation.js';
import { deserializeClipReference } from '../lib/clip-reference.js';
import { encryptAddressPassphrase, decryptAddressPassphrase } from '../lib/srt-passphrase-crypto.js';
import { getPortLease } from '../services/port-lease.js';
import { clashesAfterWrite, listenerPortRequest, resolveListenerAddress, usedListenerPorts } from '../services/listener-ports.js';

/**
 * Cross-field validation for a source's `address` given its effective
 * `streamType`. Adds a zod issue on `['address']` on failure. Shared by the
 * create and patch schemas so all stream types (including `clip`) validate
 * identically.
 *
 * `clip` sources store a serialized `ClipReference` (JSON) in `address`; it is
 * deserialized, shape-validated and semantically validated here (url → SSRF
 * checks, s3 → bucket/key, tams → 501/not-implemented) — no persisted schema
 * change (issue #275, spec: clip-story-playback.md §"Migration").
 */
function validateSourceAddress(
  streamType: string,
  address: string,
  ctx: z.RefinementCtx,
): void {
  if (streamType === 'html') {
    // html sources use a browser URL — validate for SSRF (file://, javascript:, private IPs, etc.)
    try {
      graphicUrl(address);
    } catch (err) {
      ctx.addIssue({ code: z.ZodIssueCode.custom, path: ['address'], message: err instanceof Error ? err.message : 'Invalid HTML source URL' });
    }
  } else if (streamType === 'srt' || streamType === 'efp') {
    // SRT/EFP sources: must be a valid srt:// URI pointing to a non-private host
    try {
      srtUrl(address);
    } catch (err) {
      ctx.addIssue({ code: z.ZodIssueCode.custom, path: ['address'], message: err instanceof Error ? err.message : 'Invalid SRT source address' });
    }
  } else if (streamType === 'clip') {
    // clip sources carry a serialized typed ClipReference (url/s3/tams) as JSON.
    try {
      deserializeClipReference(address);
    } catch (err) {
      ctx.addIssue({ code: z.ZodIssueCode.custom, path: ['address'], message: err instanceof Error ? err.message : 'Invalid clip reference' });
    }
  }
}

const SourceInput = z.object({
  name: z.string().min(1).max(256),
  address: z.string(),
  streamType: z.enum(['srt', 'efp', 'whip', 'html', 'clip']),
  status: z.enum(['active', 'inactive']).default('inactive'),
  liveCamera: z.boolean().optional(),
  latency: z.number().int().min(20).max(8000).optional(),
  // Optional id of the Gateway that registered this source (OL-5, spec
  // studio-gateways.md). Absent for manually-created sources; when present it
  // must name an existing gateway (validated in the route handler).
  gatewayId: z.string().optional(),
}).superRefine((data, ctx) => {
  validateSourceAddress(data.streamType, data.address, ctx);
});

const SourcePatch = z.object({
  name: z.string().min(1).max(256).optional(),
  address: z.string().optional(),
  streamType: z.enum(['srt', 'efp', 'whip', 'html', 'clip']).optional(),
  status: z.enum(['active', 'inactive']).optional(),
  liveCamera: z.boolean().optional(),
  latency: z.number().int().min(20).max(8000).optional(),
  // A gateway can re-tag or clear a source's gateway link. `null` clears the
  // link (unlinks the source); a string re-links it to an (existing) gateway.
  gatewayId: z.string().nullable().optional(),
}).superRefine((data, ctx) => {
  // Cross-field validation only applies when both address and streamType are present in
  // the same patch body. The combined-with-existing-document case is validated procedurally
  // in the route handler, which has access to the stored source document.
  if (data.address === undefined || data.streamType === undefined) {
    return;
  }
  validateSourceAddress(data.streamType, data.address, ctx);
});

/** Masks passphrase values in SRT URIs so credentials are never returned to clients. */
function maskSrtPassphrase(address: string): string {
  return address.replace(/([?&]passphrase=)[^&]*/gi, '$1***');
}

/**
 * Confirms `gatewayId` names an existing gateway. A typo would otherwise create
 * an unreachable link (the forget-gateway cascade would never find the source),
 * so an unknown id is rejected with 400. Looks the gateway up the same way the
 * gateways route does (`getGatewaysDb().get`, `src/routes/gateways.ts`).
 * Returns true when the gateway exists, false on a 404, and rethrows other DB
 * errors so they surface as a 503 like elsewhere in this module.
 */
async function gatewayExists(gatewayId: string): Promise<boolean> {
  try {
    await getGatewaysDb().get(gatewayId);
    return true;
  } catch (err: unknown) {
    if (err && typeof err === 'object' && 'statusCode' in err && (err as { statusCode: number }).statusCode === 404) {
      return false;
    }
    throw err;
  }
}

function toApi(doc: SourceDoc) {
  const { _id, _rev, type, ...rest } = doc;
  // Passphrases are stored encrypted (encv1:...); decrypt before masking so the
  // mask matches on the "passphrase=" param regardless of storage form. Legacy
  // plaintext passphrases pass through decryption unchanged.
  return { id: _id, ...rest, address: maskSrtPassphrase(decryptAddressPassphrase(rest.address)) };
}

const sourcesRoutes: FastifyPluginAsync = async (fastify) => {
  // List all sources
  fastify.get('/api/v1/sources', async (_req, reply) => {
    const db = getSourcesDb();
    let result: Awaited<ReturnType<typeof db.find>>;
    try {
      result = await db.find({ selector: { type: 'source' } });
    } catch (err) {
      fastify.log.warn({ err }, 'GET /api/v1/sources — DB query failed');
      return reply.status(503).send({ error: 'Database unavailable' });
    }
    return reply.send((Array.isArray(result?.docs) ? result.docs : []).map(toApi));
  });

  // Create a source
  fastify.post('/api/v1/sources', async (req, reply) => {
    const body = SourceInput.parse(req.body);
    // A provided gatewayId must name an existing gateway — reject a typo so the
    // source is never linked to a gateway that can never reach it (spec
    // studio-gateways.md, forget-cascade selector `{ type: 'source', gatewayId }`).
    if (body.gatewayId !== undefined && !(await gatewayExists(body.gatewayId))) {
      return reply.status(400).send({ error: `Gateway "${body.gatewayId}" not found`, statusCode: 400 });
    }
    const isSrt = body.streamType === 'srt' || body.streamType === 'efp';
    const id = `src-${randomUUID()}`;
    // A listener source binds a port on the shared Strom: it must lie inside this
    // instance's lease and no other source or output may hold it. Port 0 asks
    // for the lowest free one. Written, then checked again, because two clients
    // registering at once can both be handed the same free port.
    let used = isSrt ? await usedListenerPorts() : [];
    for (let attempt = 0; ; attempt++) {
      let address = body.address;
      let port: number | null = null;
      if (isSrt) {
        const resolved = resolveListenerAddress(body.address, getPortLease(), used);
        if (!resolved.ok) {
          return reply.status(resolved.statusCode).send({ error: resolved.error, statusCode: resolved.statusCode });
        }
        ({ address, port } = resolved);
      }
      const now = new Date().toISOString();
      const doc: SourceDoc = {
        _id: id,
        type: 'source',
        name: body.name,
        // Encrypt any embedded SRT passphrase before it touches CouchDB (issue #160).
        address: encryptAddressPassphrase(address),
        streamType: body.streamType,
        status: body.status,
        liveCamera: body.liveCamera,
        latency: body.latency,
        // Only persist gatewayId when supplied — existing clients omit it and
        // the field stays absent, keeping the doc backward-compatible.
        ...(body.gatewayId !== undefined ? { gatewayId: body.gatewayId } : {}),
        createdAt: now,
        updatedAt: now,
      };
      const written = await getSourcesDb().insert(doc);
      if (port === null) return reply.status(201).send(toApi(doc));
      used = await usedListenerPorts();
      const clash = clashesAfterWrite(used, { kind: 'source', id }, port);
      if (!clash) return reply.status(201).send(toApi(doc));
      await getSourcesDb().destroy(id, written.rev);
      if (listenerPortRequest(body.address) !== 0 || attempt >= 3) {
        return reply.status(409).send({ error: `SRT listener port ${port} is already used by ${clash.kind} "${clash.name}"`, statusCode: 409 });
      }
      fastify.log.warn({ port, clash }, 'listener port was taken while assigning it, choosing another');
    }
  });

  // Get a source
  fastify.get<{ Params: { id: string } }>('/api/v1/sources/:id', async (req, reply) => {
    try {
      const doc = await getSourcesDb().get(req.params.id);
      return reply.send(toApi(doc));
    } catch {
      return reply.status(404).send({ error: 'Source not found', statusCode: 404 });
    }
  });

  // Update a source
  fastify.patch<{ Params: { id: string } }>('/api/v1/sources/:id', async (req, reply) => {
    const body = SourcePatch.parse(req.body);
    // A non-null gatewayId in the patch must name an existing gateway (same
    // guard as create). `null` clears the link and needs no lookup.
    if (typeof body.gatewayId === 'string' && !(await gatewayExists(body.gatewayId))) {
      return reply.status(400).send({ error: `Gateway "${body.gatewayId}" not found`, statusCode: 400 });
    }
    try {
      const doc = await getSourcesDb().get(req.params.id);
      // Determine effective streamType and address after the patch. Validate
      // against the plaintext form — a new body.address is already plaintext,
      // while the stored doc.address may hold an encrypted passphrase.
      const effectiveStreamType = body.streamType ?? doc.streamType;
      const effectiveAddress = body.address ?? decryptAddressPassphrase(doc.address);
      if (effectiveAddress) {
        try {
          if (effectiveStreamType === 'html') {
            graphicUrl(effectiveAddress);
          } else if (effectiveStreamType === 'srt' || effectiveStreamType === 'efp') {
            srtUrl(effectiveAddress);
          } else if (effectiveStreamType === 'clip') {
            // clip sources carry a serialized typed ClipReference (JSON) in address.
            deserializeClipReference(effectiveAddress);
          }
        } catch (err) {
          return reply.status(400).send({ error: err instanceof Error ? err.message : 'Invalid source address' });
        }
        // Re-check the port only when the address or stream type changes — a
        // rename must not fail because an older source predates the lease. Port 0
        // keeps the port the source already has when that is still valid.
        if (
          (body.address !== undefined || body.streamType !== undefined) &&
          (effectiveStreamType === 'srt' || effectiveStreamType === 'efp')
        ) {
          const stored = doc.streamType === 'srt' || doc.streamType === 'efp' ? listenerPortRequest(doc.address) : null;
          const resolved = resolveListenerAddress(effectiveAddress, getPortLease(), await usedListenerPorts(), {
            exclude: { kind: 'source', id: doc._id },
            keep: stored,
          });
          if (!resolved.ok) {
            return reply.status(resolved.statusCode).send({ error: resolved.error, statusCode: resolved.statusCode });
          }
          if (body.address !== undefined) body.address = resolved.address;
        }
      }
      // Encrypt the passphrase in an updated address before persisting. When the
      // patch leaves the address untouched, keep the already-stored value as-is.
      const addressPatch = body.address !== undefined
        ? { address: encryptAddressPassphrase(body.address) }
        : {};
      // gatewayId is handled explicitly: a string re-links, `null` clears the
      // link (the field is removed so `findTrusted({ gatewayId })` no longer
      // matches), and an absent value leaves the stored link untouched. Strip
      // it from the spread so a raw `null` never lands in the persisted doc.
      const { gatewayId: patchGatewayId, ...bodyRest } = body;
      const updated: SourceDoc = { ...doc, ...bodyRest, ...addressPatch, updatedAt: new Date().toISOString() };
      if (patchGatewayId === null) {
        delete updated.gatewayId;
      } else if (typeof patchGatewayId === 'string') {
        updated.gatewayId = patchGatewayId;
      }
      await getSourcesDb().insert(updated);
      return reply.send(toApi(updated));
    } catch (err: unknown) {
      if (err && typeof err === 'object' && 'statusCode' in err && (err as { statusCode: number }).statusCode === 404) {
        return reply.status(404).send({ error: 'Source not found', statusCode: 404 });
      }
      throw err;
    }
  });

  // Delete a source
  fastify.delete<{ Params: { id: string } }>('/api/v1/sources/:id', async (req, reply) => {
    try {
      const doc = await getSourcesDb().get(req.params.id);

      // Block deletion if source is used by an active/activating production
      // findTrusted: operators are literals here; req.params.id is only ever a
      // scalar value inside $elemMatch, never a selector fragment (#257)
      const activeProductions = await getDb().findTrusted({
        selector: { type: 'production', status: { $in: ['active', 'activating'] }, 'sources': { $elemMatch: { sourceId: req.params.id } } },
        fields: ['_id', 'name'],
        limit: 1,
      });
      if (activeProductions.docs.length > 0) {
        const prod = activeProductions.docs[0] as unknown as Pick<ProductionDoc, '_id' | 'name'>;
        return reply.status(409).send({ error: `Source is in use by active production "${prod.name}"` });
      }

      // Remove references from inactive productions and record a warning
      // findTrusted: operators are literals here; req.params.id is only ever a
      // scalar value inside $elemMatch, never a selector fragment (#257)
      const inactiveProductions = await getDb().findTrusted({
        selector: { type: 'production', status: 'inactive', 'sources': { $elemMatch: { sourceId: req.params.id } } },
        fields: ['_id', 'name', 'sources', 'deletionWarnings'],
        limit: 100,
      });
      for (const p of inactiveProductions.docs) {
        const prod = p as unknown as ProductionDoc;
        const warnings = prod.deletionWarnings ?? [];
        warnings.push({ type: 'source', name: doc.name });
        await updateProductionDoc(prod._id, {
          sources: prod.sources.filter((s) => s.sourceId !== req.params.id),
          deletionWarnings: warnings,
        });
      }

      await getSourcesDb().destroy(doc._id, doc._rev!);
      return reply.status(204).send();
    } catch (err: unknown) {
      if (err && typeof err === 'object' && 'statusCode' in err && (err as { statusCode: number }).statusCode === 404) {
        return reply.status(404).send({ error: 'Source not found' });
      }
      throw err;
    }
  });
};

export default sourcesRoutes;